EURO removed from list of languages
I don't know if the language list is generated automatically from the language files, but the currency conversion EURO file is showing up in the list of languages. Additionally the eu language code belongs to Euskara (Basque language), so as a minor change I have removed the eu lang from the list.