Improve decimals and add case and gender to cardinal
Decimals are now converted to a more usual form with proper names of the
fractions and with the conjuction "i" instead of the word "przecinek"
for the decimal comma.

to_cardinal() now supports, to a limited extent, the gender and case parameters.
Use gender="f" for "feminine". Any other value returns the inanimate masculine
forms, which are identical to neuter forms. Only "nominative" and
"genitive" cases are implemented.